Move RXF to examples dir
