+++ examples/rxf/src/__init__.py
@@ -1,0 +1,1 @@
+from . import label_smoothed_cross_entropy_r3f, sentence_prediction_r3f
